refactor transform path ops helpers
